+++ jinja2xlsx/render.py
@@ -1,5 +1,4 @@
 from dataclasses import dataclass
-from itertools import tee
 from typing import Tuple
 
 from openpyxl import Workbook
@@ -20,9 +19,9 @@ class Renderer:
     def __call__(self) -> Workbook:
         wb = Workbook()
 
-        cells_to_fill, cells_to_style = tee(self._generate_cells(wb.active))
-        self._fill_cells(cells_to_fill)
-        self._style_cells(cells_to_style)
+        cells = list(self._generate_cells(wb.active))
+        self._fill_cells(cells)
+        self._style_cells(cells)
 
         adjuster = Adjuster(wb.active)
         adjuster.adjust_columns(self.parser.columns)